+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.1] - 03.11.2025
 
 - **Szyfr AES ECB** - "poprawna" implementacja szyfru AES w trybie ECB
+- **Szyfr ChaCha20** = dodanie szyfru ChaCha20
 
 ## [1.0.5] - 28.10.2025